basic app is now working properly - can view prices of bitcoin in all currencies now that the data models are being parse from JSON, into BitcoinCurrencyModel and persisted via Realm

+++ app/src/main/java/bacon/john/bitcoinpricewatch/bitcoin/BitcoinCurrencyModel.java
@@ -15,7 +15,7 @@ public class BitcoinCurrencyModel extends RealmObject {
     private double bid;
     private double last;
 
-    private int updatedAt;
+    private long updatedAt;
 
     public String getCode() {
         return code;
@@ -39,10 +39,10 @@ public class BitcoinCurrencyModel extends RealmObject {
         return last;
     }
     public void setLast(double last) { this.last = last; }
-    public int getUpdatedAt() {
+    public long getUpdatedAt() {
         return updatedAt;
     }
-    public void setUpdatedAt(int updatedAt) {
+    public void setUpdatedAt(long updatedAt) {
         this.updatedAt = updatedAt;
     }
 }
